+++ lib/ui/wellness/WellnessHomePanel.dart
@@ -23,6 +23,7 @@ import 'package:illinois/service/Guide.dart';
 import 'package:illinois/ui/WebPanel.dart';
 import 'package:illinois/ui/guide/GuideDetailPanel.dart';
 import 'package:illinois/ui/wellness/WellnessHealthScreenerWidgets.dart';
+import 'package:illinois/ui/wellness/WellnessMentalHealthContentWidget.dart';
 import 'package:illinois/ui/wellness/WellnessResourcesContentWidget.dart';
 import 'package:illinois/ui/wellness/appointments/WellnessAppointmentsHomeContentWidget.dart';
 import 'package:illinois/ui/wellness/rings/WellnessRingsHomeContentWidget.dart';
@@ -40,7 +41,7 @@ import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 import 'package:url_launcher/url_launcher.dart';
 
-enum WellnessContent { dailyTips, rings, todo, appointments, healthScreener, podcast, resources, struggling }
+enum WellnessContent { dailyTips, rings, todo, appointments, healthScreener, podcast, resources, struggling, mentalHealth }
 
 class WellnessHomePanel extends StatefulWidget {
   final WellnessContent? content;
@@ -215,6 +216,8 @@ class _WellnessHomePanelState extends State<WellnessHomePanel>
       return WellnessContent.podcast;
     } else if (code == 'resources') {
       return WellnessContent.resources;
+    } else if (code == 'mental_health') {
+      return WellnessContent.mentalHealth;
     } else if (code == 'struggling') {
       return WellnessContent.struggling;
     } else {
@@ -255,6 +258,8 @@ class _WellnessHomePanelState extends State<WellnessHomePanel>
         return (guideId != null) ? GuideDetailWidget(key: _podcastKey, guideEntryId: guideId, headingColor: Styles().colors?.background) : Container();
       case WellnessContent.resources:
         return WellnessResourcesContentWidget();
+      case WellnessContent.mentalHealth:
+        return WellnessMentalHealthContentWidget();
       case WellnessContent.struggling:
         String? guideId = _loadWellcomeResourceGuideId('where_to_start');
         return (guideId != null) ? GuideDetailWidget(key: _strugglingKey, guideEntryId: guideId, headingColor: Styles().colors?.background) : Container();
@@ -316,6 +321,8 @@ class _WellnessHomePanelState extends State<WellnessHomePanel>
         return _loadContentString('panel.wellness.section.screener.label', 'Illinois Health Screener');
       case WellnessContent.resources:
         return _loadContentString('panel.wellness.section.resources.label', 'Wellness Resources', language: language);
+      case WellnessContent.mentalHealth:
+        return _loadContentString('panel.wellness.section.mental_health.label', 'Mental Health', language: language);
       case WellnessContent.podcast:
         return _loadContentString('panel.wellness.section.podcast.label', 'Healthy Illini Podcast', language: language);
       case WellnessContent.struggling: